Refactor Maki.
