+++ app/models/User.js
@@ -20,53 +20,4 @@ UserSchema.plugin( slug( 'username' , {
   required: true
 }) );
 
-UserSchema.pre('save', function (next) {
-  this.meta = {};
-
-  this.meta.wasNew      = this.isNew;
-  this.meta.wasModified = this.isModified();
-
-  // test
-  if (this.meta.wasModified) {
-    this.meta.modifiedPaths = this.modifiedPaths();
-    this.meta.old = this.toObject();
-  }
-
-  next();
-})
-
-UserSchema.post('save', function(doc) {
-  var path = '/people';
-
-  if (doc.meta.wasNew) {
-    console.log('WAS NEW');
-
-    // stub the pre-condition as an empty set
-    var collection = [];
-    // observe the empty set
-    var observer = patch.observe( collection );
-    // add our "new" doc
-    collection.push( doc );
-    // generate our patch set from the stub
-    var patches = patch.generate( observer );
-    // publish the patch set
-    app.redis.publish( path , JSON.stringify(patches) );
-  } else if (doc.meta.wasModified) {
-    console.log('MODIFIED:' , doc.meta.modifiedPaths );
-
-    var observer = patch.observe( doc.meta.old );
-
-    // generate our patch set from the doc
-    var patches = patch.generate( observer );
-    // publish the patch set
-    app.redis.publish( path , JSON.stringify(patches) );
-
-  }
-});
-
-var User = mongoose.model('User', UserSchema);
-
-// export the model to anything requiring it.
-module.exports = {
-  User: User
-};
+module.exports = UserSchema;
