+++ package.json
@@ -21,30 +21,33 @@
   "readmeFilename": "README.md",
   "gitHead": "22dbaa6e09ffb728b79cecf128f4694fe12d5661",
   "dependencies": {
-    "passport": "~0.1.16",
-    "jade": "~0.28.2",
-    "passport-local-mongoose": "~0.2.1",
-    "express": "~4.3.2",
-    "mongoose": "~3.8.1",
-    "passport-local": "~0.1.6",
-    "less-middleware": "~0.1.14",
-    "mongoose-slug": "^1.3.0",
-    "flashify": "~0.1.2",
-    "underscore": "~1.6.0",
-    "async": "~0.7.0",
-    "cookie-parser": "~1.1.0",
-    "body-parser": "~1.2.2",
-    "express-session": "~1.2.1",
     "asset-rack": "git://github.com/martindale/asset-rack.git#8d03859922d83f4839086b1208ff6aab28a0a272",
-    "ws": "~0.4.31",
-    "hbs": "~2.7.0",
-    "provide": "^0.1.3",
-    "path-to-regexp": "~0.2.1",
-    "redis": "~0.10.3",
+    "async": "~0.7.0",
+    "body-parser": "~1.2.2",
+    "connect-assets": "^4.3.0",
     "connect-redis": "~2.0.0",
-    "node-uuid": "^1.4.1",
+    "cookie-parser": "~1.1.0",
+    "express": "~4.3.2",
+    "express-session": "~1.2.1",
+    "fast-json-patch": "^0.3.6",
+    "flashify": "~0.1.2",
+    "hbs": "~2.7.0",
+    "jade": "~0.28.2",
     "json-patch": "^0.4.3",
-    "fast-json-patch": "^0.3.6"
+    "less-middleware": "~0.1.14",
+    "moment": "^2.8.1",
+    "mongoose": "~3.8.1",
+    "mongoose-slug": "^1.3.0",
+    "node-uuid": "^1.4.1",
+    "passport": "~0.1.16",
+    "passport-local": "~0.1.6",
+    "passport-local-mongoose": "~0.2.1",
+    "path-to-regexp": "~0.2.1",
+    "preston": "^0.2.1",
+    "provide": "^0.1.3",
+    "redis": "~0.10.3",
+    "underscore": "~1.6.0",
+    "ws": "~0.4.31"
   },
   "devDependencies": {
     "assert": "~1.1.1",
